refactor(cowork-voicel): 공통 패키지 경로 평탄화
